feat(2 Players): Yay I spent way longer than I care to admit to debug this. Found a funny quirk where when defining the indices for the rendering you need to define that the slice is &[u16] likely cause by default Rust would think it's &[u32].
I spent a few hours trying to debug this and then accidentally found it by just A/B testing
with code I knew to work and changing one thing at a time.

In hindsight I am kinda frustrated it took that long but oh well.
The next challenges are in no particular order:
- Get the ball moving
- Allow simultaneous movement
- Get the scores
- Add the midline thingey
- Play?