feat(bounds + simultaneous movement): simplified the code slightly so that I don't need as many functions that do the same thing and I just pass references instead.
But now I have a weird bug where sometimes it's possible to break through the bounds if
simultaneous movement is attempted.

I think this has something to do with the buffering system or the callback.
But yeah to investigate. For now it works good enough that I think I can focus on making this a game
and adding ball movement!